Prototype of HelperMode
In my push, I implemented function that show the range of explosion, AKA helper mode, by default it is on.

There are some code for helper mode to show on the main menu but did not finish yet.

Because the mode that Spritebatch mode set, it is ignoring the depth of pictures so the predict range can only show above the player. If we want to fix that, I think give everything a depth would be helpful.